Add error handling on Action menu and common code refactoring
